Fix district baking for single state
Fix a small bug in the `bake_geography` management command that causes
it to fail when the user specifies individual state FIPS codes for
states with multiple districts. This bug did not affect baking the
entire country by specifying the FIPS code `00` and did not affect
baking individual states with a single at-large district.